+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.17.1-dev
 
 * Require Dart 2.18
+* Improve docs for `splitAfter` and `splitBefore`.
 
 ## 1.17.0